% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndividualFunctions.R
\name{crossIndividuals}
\alias{crossIndividuals}
\title{Cross two indiviiduals
Workhorse function to cross two individuals. Usually isn't called directly but useful for customized simulations
Takes two parents, recombines them, than randomly generates a progeny by a combination of one haplotype of each.}
\usage{
crossIndividuals(parent1, parent2, genome, incompatibility = NULL,
  sex = "random", Nrecombs = 1)
}
\arguments{
\item{parent1}{first parent. an individual object}

\item{parent2}{second parent. an indiviudal object}

\item{genome}{a genome object, should correspond to both individuals}

\item{incompatibility}{a list of incompatibilities. Each incompatibility is a list with two elements: params and func. The func is the function that tests the recombined parents and the offspring and determines if it's dead or not. See the functions zeelPeelSelection and medeaSelection for example of functions included in the package.}

\item{sex}{what sex should the offspring have. either "random", or male (0,"0","male","m") or female (1,"1","female","f")}

\item{Nrecombs}{number of recombination per chromosome}
}
\description{
Cross two indiviiduals
Workhorse function to cross two individuals. Usually isn't called directly but useful for customized simulations
Takes two parents, recombines them, than randomly generates a progeny by a combination of one haplotype of each.
}
