% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{findProxyForPosition}
\alias{findProxyForPosition}
\title{Find closest marker to a position}
\usage{
findProxyForPosition(genome, chrom, pos)
}
\arguments{
\item{genome}{A genome object to search in}

\item{chrom}{The chromosome to query}

\item{pos}{position in the chromosome}
}
\value{
the name of the closest variant
}
\description{
Find closest marker to a position
}
