% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndividualFunctions.R
\name{getGenotypes}
\alias{getGenotypes}
\title{Get the genotype of the individual at a variant or position
Either an exact chromosome and position or a variant name are required
see the function findProxyForPosition for an easy way to find closest variant to any position in the genome}
\usage{
getGenotypes(individual, genome, haplotype = c(1, 2), var = NULL,
  chrom = NULL, pos = NULL)
}
\arguments{
\item{individual}{an individual to query}

\item{genome}{a genome object corresponding to the individual}

\item{haplotype}{which haplotype to query}

\item{var}{variant name}

\item{chrom}{chromosome to query}

\item{pos}{position to query}
}
\description{
Get the genotype of the individual at a variant or position
Either an exact chromosome and position or a variant name are required
see the function findProxyForPosition for an easy way to find closest variant to any position in the genome
}
