% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkPopSim.R
\name{getIndividualMatingProbabilities}
\alias{getIndividualMatingProbabilities}
\title{Return a relative measure of mating probability for a list of individuals
This function shouldn't usually be called directly, use getMatingSubpopulation instead}
\usage{
getIndividualMatingProbabilities(individuals, lociList, genome)
}
\arguments{
\item{individuals}{a list of individuals}

\item{lociList}{see fitnessLoci parameter in getMatingSubpopulation}

\item{genome}{a genome object}
}
\description{
Return a relative measure of mating probability for a list of individuals
This function shouldn't usually be called directly, use getMatingSubpopulation instead
}
