% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkPopSim.R
\name{getMatingSubpopulation}
\alias{getMatingSubpopulation}
\title{Use the fitness loci to choose the submating population out of the pool of males or females.}
\usage{
getMatingSubpopulation(genome, pop, matingFrac, fitnessLoci = NULL,
  sex = c("male", "female"))
}
\arguments{
\item{genome}{A genome object generated with the package}

\item{pop}{A list of individuals of males and females to pick from}

\item{matingFrac}{the fraction of the relevant sex population which will be chosen. between 0 and 1 (1 is no selection)}

\item{fitnessLoci}{A data frame with the following columns:
chrom: chromosome
marker: name of the marker to use (see findProxyForPosition for easy way to get marker for locus of interest)
direction: (0,1) which allele confers an advantage?
effect: the effect size of the locus. How much advantage does it confer?}

\item{sex}{choose a population of males or females?}
}
\description{
Use the fitness loci to choose the submating population out of the pool of males or females.
}
