% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndividualFunctions.R
\name{is.individual}
\alias{is.individual}
\title{Test if x is in an individual object}
\usage{
is.individual(x)
}
\arguments{
\item{x}{an object to test}
}
\description{
Test if x is in an individual object
}
