% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{N2xCB4856.genome}
\alias{N2xCB4856.genome}
\title{A genome object for N2 and CB4856}
\format{An object of class \code{environment} of length 14.}
\source{
\url{https://github.com/AndersenLab/N2xCB4856-RIAILS/tree/master/dataes}
}
\usage{
N2xCB4856.genome
}
\description{
An example genome project, built from the C. elegans strains N2 and CB4856 based on the markers from Rockman and Kruglyak (2009), doi.org/10.1371/journal.pgen.1000419
}
\keyword{datasets}
