% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeFunctions.R
\name{newGenome}
\alias{newGenome}
\title{Create a new genome object - object including all markers and the recombination rates between them}
\usage{
newGenome(name, Nchrom, markerNames, markerChrom, markerPos, recombProbs, map,
  Nrecs = 1e+06)
}
\arguments{
\item{name}{Name of the genome object}

\item{Nchrom}{Number of chromosomes}

\item{markerNames}{A vector of the name of the markers}

\item{markerChrom}{A vector of the chromosome of each marker}

\item{markerPos}{A vector of the position (in basepairs) of each marker}

\item{recombProbs}{A vector of the recombination probability between each two successive markers (can be generated from a genetic map using recombProbFromGeneticDistance)}

\item{map}{The genetic distance between two successive genetic markers}

\item{Nrecs}{Number of recombination positions to simulate when generating the genome object (recombination spots in crosses will be cycled through this object to save time. Defaults to 1,000,000)}
}
\description{
Create a new genome object - object including all markers and the recombination rates between them
}
