% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseAndPlot.R
\name{parseAllGenotypes}
\alias{parseAllGenotypes}
\title{Function to parse genotype table from a population of individuals and genome object
Useful for generating data frames to be used for plotting}
\usage{
parseAllGenotypes(pop, genome, summarize = F)
}
\arguments{
\item{pop}{a population of individuals}

\item{genome}{a genome object}

\item{summarize}{If FALSE, output raw genotypes (big matrix). TRUE: Output allele frequencies}
}
\description{
Function to parse genotype table from a population of individuals and genome object
Useful for generating data frames to be used for plotting
}
