% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkPopSim.R
\name{performMating}
\alias{performMating}
\title{Mate a population of males to a population of females.}
\usage{
performMating(males, females, genome, offPerMating, ncores = 1,
  Nrecombs = 1, incompatibility = NULL)
}
\arguments{
\item{males}{a list of male individuals}

\item{females}{a list of female individuals}

\item{genome}{A genome object generated with the package}

\item{offPerMating}{how many offspring does each female generate}

\item{ncores}{Number of cores to be used. If more than one core is specified, a cluster object will be created and used.}

\item{Nrecombs}{the number of recombinations per chromosome per cross (defaults to 1)}

\item{incompatibility}{a list of incompatibilities. Each incompatibility is a list with two elements: params and func. The func is the function that tests the recombined parents and the offspring and determines if it's dead or not. See the functions zeelPeelSelection and medeaSelection for example of functions included in the package.}
}
\description{
Mate a population of males to a population of females.
}
