% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseAndPlot.R
\name{plotGenotypeAndMarkerList}
\alias{plotGenotypeAndMarkerList}
\title{Plot a list of genotypes and markers
An extension of plotGenotypeAndMarkerTable to plot multiple populations/generations.}
\usage{
plotGenotypeAndMarkerList(genotypeAndMarkerList, calcFreqs = T,
  chromPos = NULL, markers = NULL, internal = F)
}
\arguments{
\item{genotypeAndMarkerList}{a list where each element is a genotypeAndMarkerTable}

\item{calcFreqs}{Do allele frequencies need to be calculated from the table (see summarize in parseAllGenotypes function)}

\item{chromPos}{optional matrix with two columns (chromosome and position). If given, only these positions will be plotted}

\item{markers}{optional vector of variants. if given, only these variants will be plotted.}

\item{internal}{generate plot (FALSE) or just return plotting dataframe (TRUE)?}
}
\description{
Plot a list of genotypes and markers
An extension of plotGenotypeAndMarkerTable to plot multiple populations/generations.
}
