% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseAndPlot.R
\name{plotGenotypeAndMarkerTable}
\alias{plotGenotypeAndMarkerTable}
\title{Plot genotype and marker table}
\usage{
plotGenotypeAndMarkerTable(genotypeAndMarkerTable, calcFreqs = T,
  internal = F)
}
\arguments{
\item{genotypeAndMarkerTable}{a table generated with parseAllGenotypes}

\item{calcFreqs}{Do allele frequencies need to be calculated from the table (see summarize in parseAllGenotypes function)}

\item{internal}{generate plot (FALSE) or just return plotting dataframe (TRUE)?}
}
\description{
Plot genotype and marker table
}
