% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeFunctions.R
\name{recombProbFromGeneticDistance}
\alias{recombProbFromGeneticDistance}
\title{Convert genetic distance to probability of recombination}
\usage{
recombProbFromGeneticDistance(distances)
}
\arguments{
\item{distances}{A vector of genetic distances between the markers}
}
\description{
Convert genetic distance to probability of recombination
}
\note{
This function assumes a single recombination event per chromosome!!!
}
