% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndividualFunctions.R
\name{selfHermaphrodite}
\alias{selfHermaphrodite}
\title{Perform a selfing
Hasn't been thoroughly tested, basically generates a male version of the same individuals (with random haplotype chosen to be shortened) and crosses them}
\usage{
selfHermaphrodite(parent, genome, incompatibility = NULL, sex = "female",
  Nrecombs = 1)
}
\arguments{
\item{parent}{the individual to self}

\item{genome}{a genome object. Corresponds to the individual}

\item{incompatibility}{a list of incompatibilities. Each incompatibility is a list with two elements: params and func. The func is the function that tests the recombined parents and the offspring and determines if it's dead or not. See the functions zeelPeelSelection and medeaSelection for example of functions included in the package.}

\item{sex}{the sex of the offspring}

\item{Nrecombs}{number of recombinations per chromosome}
}
\description{
Perform a selfing
Hasn't been thoroughly tested, basically generates a male version of the same individuals (with random haplotype chosen to be shortened) and crosses them
}
