% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGenome.R
\name{expandGenomeToVariantlist}
\alias{expandGenomeToVariantlist}
\title{The function expands a}
\usage{
expandGenomeToVariantlist(genome, variantNames = NULL, variantChrom,
  variantPos, matchNames = F)
}
\arguments{
\item{genome}{a genome object}

\item{variantNames}{vector of variant names. If not given will be generated as chrom:pos}

\item{variantChrom}{vector of variant chromsomes. Names must match genome object}

\item{variantPos}{vector of variant positions.}

\item{matchNames}{boolean: If true, the function will look for the new variants by position in the old genome and carry over the names to the new genome.}
}
\description{
The function expands a
}

