% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGenome.R
\name{recombProbFromGeneticDistance}
\alias{recombProbFromGeneticDistance}
\title{This function assumes a single recombination event per chromosome!!!}
\usage{
recombProbFromGeneticDistance(distances)
}
\arguments{
\item{distances}{A vector of genetic distances between the markers}
}
\description{
This function assumes a single recombination event per chromosome!!!
}

